machine-learning/mcp-rag/python/pyproject.toml
[project]
name = "mcp-rag"
version = "0.1.0"
description = ""
authors = [
    {name = "Alexander Foley",email = "alexandergfoley@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0.0"
dependencies = [
    "fastmcp (>=2.14.1,<3.0.0)",
    "langchain (>=1.2.0,<2.0.0)",
    "langchain-community (>=0.4.1,<0.5.0)",
    "langchain-ollama (>=1.0.1,<2.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "langchain-postgres (>=0.0.16,<0.0.17)",
    "starlette (>=0.50.0,<0.51.0)",
    "uvicorn (>=0.39.0,<0.40.0)",
    "fastapi (>=0.126.0,<0.127.0)",
    "sentence-transformers (>=5.2.0,<6.0.0)",
    "pypdf2 (>=3.0.1,<4.0.0)",
    "pymupdf (>=1.26.7,<2.0.0)",
    "beautifulsoup4 (>=4.14.3,<5.0.0)",
    "lxml (>=6.0.2,<7.0.0)",
    "markdown (>=3.10,<4.0)",
    "tqdm (>=4.67.1,<5.0.0)",
    "pydantic (>=2.12.5,<3.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
black = "^25.12.0"
isort = "^7.0.0"
